refactor: extract FontSupport module from TextBubble and MonitorRenderer
Extract duplicated `find_font` and `fill_circle` methods into a shared
`Render::FontSupport` module included by both TextBubble and
MonitorRenderer.

Refs #426
